+++ docs/whats-new.rst
@@ -52,6 +52,16 @@ We added the Ocean-OPS (former JCOMMOPS) dashboard for all floats and the Argo-B
     # or
     argopy.dashboard(5904797, 12, type='bgc')
 
+- New utility function :class:`argopy.utilities.ArgoNVSReferenceTables` to retrieve Argo Reference Tables. (:commit:`cc8fdbe132874b71b35203053626cc29ae7d19c4`) by `G. Maze <http://www.github.com/gmaze>`_.
+
+.. code-block:: python
+
+    from argopy.utilities import ArgoNVSReferenceTables
+    R = ArgoNVSReferenceTables()
+    R.all_tbl_name()
+    R.tbl(3)
+    R.tbl('R09')
+
 **Internals**
 
 - Fix bug in erddap fata fetcher that was causing a `profile` request to do not account for cycle numbers. (:commit:`301e557fdec1f2d536841464b383edc3a4c4a62d`) by `G. Maze <http://www.github.com/gmaze>`_.
@@ -61,6 +71,8 @@ We added the Ocean-OPS (former JCOMMOPS) dashboard for all floats and the Argo-B
 **Breaking changes with previous versions**
 
 - Remove deprecation warnings for: 'plotters.plot_dac', 'plotters.plot_profilerType'. These now raise an error.
+- ``argopy.plotters`` module replaced by ``argopy.plot``
+
 
 v0.1.10 (4 Mar. 2022)
 ---------------------